Unified boundary. Proportional area point select. Interior point gen.
Boundary is now unified and stored in .txt and .npy files (ignored in commits due to size).

Training selects points from preloaded data. Number of points is proportional to the surface areas of the different components of the boundary.

Coordinates for interior points can now be generated by a function. These will be used in training along with boundary points, but only tested against Maxwell's equations, not MSE (because we don't have measurements there).